refactor(rsc): rename lazy-client-css to lazy-css
The directory now contains tests for both client and server lazy CSS cases.
