+++ packages/plugin-rsc/examples/basic/src/routes/lazy-css/client-direct.css
