+++ packages/plugin-rsc/examples/basic/src/routes/lazy-css/lazy-server-dep.tsx
