+++ packages/plugin-rsc/examples/basic/src/routes/lazy-css/server.tsx
