+++ packages/plugin-rsc/examples/basic/src/routes/root.tsx
@@ -34,11 +34,8 @@ import { TestHmrSwitchClient } from './hmr-switch/client'
 import { TestHmrSwitchServer } from './hmr-switch/server'
 import { TestHydrationMismatch } from './hydration-mismatch/server'
 import { TestImportMetaGlob } from './import-meta-glob/server'
-import { TestLazyClientCss } from './lazy-client-css/client'
-import {
-  TestLazyClientCssServer,
-  TestLazyServerCss,
-} from './lazy-client-css/server'
+import { TestLazyClientCss } from './lazy-css/client'
+import { TestLazyClientCssServer, TestLazyServerCss } from './lazy-css/server'
 import { TestModuleInvalidationServer } from './module-invalidation/server'
 import { TestPayloadServer } from './payload/server'
 import { TestReactCache } from './react-cache/server'
